Copy with MIME type text/uri-list
This backwards-compatible change allows for a huge benefit: other apps (file managers, chat apps, etc.) might understand that the contents of the clipboard is not just any text, but (a list of) files.
This enables for scenarios such as:

- Find a file/picture that you want to send to a friend in a chat app, copy using xplr, paste in chat app - and pictures would be sent, instead of your file names.
- Copy files in xplr and paste in another file browser.

Not apps support this, obviously, so for them the behavior would be exactly as today - they'd just read the clipboard as if it was plain text.

There's no change for xplr pasting capabilities as it doesn't care of the MIME type, but in principle it _could_ check whether clipboard type is text/uri-list and early abort if it's not (I believe other file managers like Nemo do exactly this, although Nemo devised their own custom MIME type in addition to text/uri-list...)